Feat: fetch invitations with react query
